refactor: extract snapshot unzip/metadata logic into SnapshotFixture
Version88Test and the upcoming Version88GoldenTest both duplicated the
unzip-to-tempdir + read-metadata.json setup. Extracted into a shared
SnapshotFixture record (unzip(zipPath, version) + cleanup()) in io.zell.zdb
so the v8.1-v8.7 migration can reuse it instead of re-copying the block.
